Boiler plate / Web sockets
I have attempted this project before using firebase as a way to keep the videos in sync, i did however run into issues trying to sync the videos in realtime due to the delay in reading and writing to the db. Im hoping i can get around that using websockets to have a backend that keeps all clients in sync to make this work. For now im just trying to get used to using express and sockets and once i have something resembling a synched video i will move on to actually building out the app.

At the moment all this does is print some simple numbers to the front end and for some reason it doesnt work. I believe its due to cors but im not sure yet.