Connecting the front and back end.
I have added the socket.io library here to connect my front and back end together. Yesterday it was not working. I realised this was because of both cors and the fact that i was listening on port 3000 for my frontend but in my case its actually on port 5173. So a simple fix thankfully.