Helper Functions
I have called them helpers but really these functions will be the building blocks of this application. I have tried to cover everything i think that the user will need to be able to do and encapsulated all the logic for it in the room.js and user.js files. These functions will be responsible for controlling the functionality of our backend. No doubt i will have missed some but i like to get something down first to have somewhere to start. At this moment i am lacking the knowledge to fully implement some of these methods that require the use of socket io to emit and receive events. I will learn what i can about it and implement here as i go along.